fix(dashboard): make tsc generate render model types again
