+++ dashboard/src/contexts/content-model-context.ts
@@ -1,8 +1,9 @@
 import React from 'react';
 import { ContentModelInstance } from '~/dashboard-editor/model/content';
-import { ContentRenderModelInstance, type IContentRenderModel } from '~/dashboard-render/model/content';
+import { ContentRenderModelInstance } from '~/dashboard-render/model/content';
+import { IContentRenderModel } from '../dashboard-render/model/types';
 
-const ContentModelContext = React.createContext<ContentModelInstance | ContentRenderModelInstance | null>(null);
+export const ContentModelContext = React.createContext<ContentModelInstance | ContentRenderModelInstance | null>(null);
 
 export const ContentModelContextProvider = ContentModelContext.Provider;
