+++ dashboard/src/contexts/dashboard-context.ts
@@ -15,4 +15,3 @@ export function useDashboardContext<T = DashboardModelInstance>() {
 }
 
 export const useEditDashboardContext = () => useDashboardContext<DashboardModelInstance>();
-export const useRenderDashboardContext = () => useDashboardContext<DashboardRenderModelInstance>();
