+++ dashboard/src/contexts/index.ts
@@ -5,3 +5,4 @@ export * from './layout-state-context';
 export * from './panel-context';
 export * from './full-screen-panel-context';
 export * from './dates-provider';
+export * from './render-content-model-context';
