+++ dashboard/src/contexts/render-content-model-context.ts
@@ -1,0 +1,6 @@
+import { IContentRenderModel } from '../dashboard-render';
+import { useDashboardContext } from './dashboard-context';
+
+export const useRenderDashboardContext = () => useDashboardContext<IContentRenderModel>();
+
+// use a separate file to allow tsc generate proper types
