+++ dashboard/src/dashboard-render/model/index.ts
@@ -1,2 +1,3 @@
 export * from './dashboard';
 export * from './content';
+export { type IContentRenderModel } from './types';
