+++ dashboard/src/model/render-model/dashboard/content/panels/index.ts
@@ -1,2 +1,3 @@
 export * from './panel';
 export * from './panels';
+export * from './types';
