+++ dashboard/src/model/render-model/dashboard/content/panels/panel.ts
@@ -1,22 +1,8 @@
-import _ from 'lodash';
-import {
-  getRoot,
-  Instance,
-  SnapshotIn,
-  IMSTArray,
-  ISimpleType,
-  IStateTreeNode,
-  IOptionalIType,
-  IArrayType,
-} from 'mobx-state-tree';
-import { type ReactNode } from 'react';
+import { getRoot, Instance, SnapshotIn } from 'mobx-state-tree';
 import { TableVizComponent } from '~/components/plugins/viz-components/table';
-import { PanelMeta, type IPanelMeta } from '~/model/meta-model/dashboard/content/panel';
-import { type IPanelStyleMeta } from '~/model/meta-model/dashboard/content/panel/style';
-import { type IPanelTitleMeta } from '~/model/meta-model/dashboard/content/panel/title';
-import { type IVariableMeta } from '~/model/meta-model/dashboard/content/panel/variable';
-import { QueryRenderModelInstance, type IQueryRenderModel } from '../queries';
-import { downloadJSON } from '~/utils/download';
+import { CURRENT_SCHEMA_VERSION } from '~/model/meta-model';
+import { PanelMeta } from '~/model/meta-model/dashboard/content/panel';
+import { typeAssert } from '~/types/utils';
 import {
   formatAggregatedValue,
   getAggregatedValue,
@@ -24,9 +10,9 @@ import {
   ITemplateVariable,
   variablesToStrings,
 } from '~/utils';
-import { CURRENT_SCHEMA_VERSION, type IQueryMeta } from '~/model/meta-model';
-import { typeAssert } from '~/types/utils';
-import { DataSourceType } from '~/model/meta-model/dashboard/content/query/types';
+import { downloadJSON } from '~/utils/download';
+import { QueryRenderModelInstance } from '../queries';
+import { IPanelRenderModel } from './types';
 
 export type VariableValueMap = Record<string, string | number>;
 export type VariableAggValueMap = Record<string, string | number>;
@@ -191,38 +177,6 @@ export function getNewPanel(id: string): PanelRenderModelSnapshotIn {
       },
     },
   };
-}
-
-export interface IPanelRenderModel extends IPanelMeta {
-  readonly contentModel: {
-    content: {
-      queries: {
-        findByIDSet(ids: Set<string>): QueryRenderModelInstance[];
-      };
-    };
-  };
-  readonly queries: QueryRenderModelInstance[];
-  readonly firstQuery: QueryRenderModelInstance | null;
-  readonly firstQueryData: Array<string[] | number[] | Record<string, unknown>>;
-  readonly data: TPanelData;
-  readonly variableStrings: Record<string, ReactNode>;
-  readonly variableAggValueMap: VariableAggValueMap;
-  readonly variableValueMap: VariableValueMap;
-  readonly variableStyleMap: VariableStyleMap;
-  readonly dataLoading: boolean;
-  readonly queryStateMessages: string;
-  readonly queryErrors: string[];
-  readonly canRenderViz: boolean;
-
-  queryByID(queryID: string): IQueryRenderModel | undefined;
-  refreshData(): void;
-  downloadData(): void;
-  getSchema(): {
-    panel: IPanelMeta['json'];
-    queries: Array<IQueryRenderModel['json']>;
-    layouts: unknown;
-  };
-  downloadSchema(): void;
 }
 
 typeAssert.shouldExtends<IPanelRenderModel, Instance<typeof PanelRenderModel>>();
